# src/nemo/apigateway-service/swagger/openapi-consolidated.yaml
openapi: 3.0.3
info:
  title: AgentStudio Services API
  version: 1.0.0
  description: Consolidated API documentation for all AgentStudio services (Config
    Service)
servers:
- url: http://localhost:8080
  description: AgentStudio Gateway (local)
- url: https://agentstudio.local:8443
  description: AgentStudio Gateway (local development)
security:
- bearerAuth: []
- oauth2:
  - openid
  - profile
  - email
paths:
...
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests:
    get:
      summary: List all manifests for a dataset
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List of manifests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataSetManifest'
    post:
      summary: Create a new manifest for a dataset
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                uris:
                  type: array
                  items:
                    type: string
                metadata:
                  type: object
                schema:
                  type: object
                  description: Optional schema definition for the manifest
      responses:
        '201':
          description: Manifest created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetManifest'
        '400':
          description: Bad request
        '404':
          description: Dataset not found
        '409':
          description: Draft manifest already exists for this dataset
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/from/{sourceManifestId}:
    post:
      summary: Create a new manifest from another manifest
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: sourceManifestId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                targetDataSetId:
                  type: string
                  format: uuid
                  description: Target dataset ID (defaults to current dataset)
                metadata:
                  type: object
                  description: Metadata for the new manifest (defaults to source manifest's
                    metadata)
                schema:
                  type: object
                  description: Optional schema definition for the manifest (defaults
                    to source manifest's schema)
      responses:
        '201':
          description: Manifest created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetManifest'
        '400':
          description: Bad request
        '404':
          description: Source manifest or dataset not found
        '409':
          description: Draft manifest already exists for target dataset
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/{id}:
    get:
      summary: Get manifest by ID
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: ^[0-9a-z]{8,12}$
          description: Workspace ID (Base36, 8-12 characters)
      responses:
        '200':
          description: Manifest details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetManifest'
        '404':
          description: Manifest not found
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/{id}/files:
    patch:
      summary: Add files to a manifest (draft only) - returns pre-signed URLs
      tags:
      - Manifests
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - fileNames
              properties:
                fileNames:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Returns array of file info with pre-signed URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    preSignedUrl:
                      type: string
                    fileName:
                      type: string
        '400':
          description: Bad request or manifest not in draft status
        '404':
          description: Manifest not found
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/{id}/status:
    put:
      summary: Update manifest status
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: ^[0-9a-z]{8,12}$
          description: Workspace ID (Base36, 8-12 characters)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  type: string
                  enum:
                  - draft
                  - committed
                  - deprecated
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetManifest'
        '400':
          description: Bad request
        '404':
          description: Manifest not found
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/{id}/metadata:
    put:
      summary: Update manifest metadata (draft only)
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: ^[0-9a-z]{8,12}$
          description: Workspace ID (Base36, 8-12 characters)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - metadata
              properties:
                metadata:
                  type: object
      responses:
        '200':
          description: Metadata updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetManifest'
        '400':
          description: Bad request or manifest not in draft status
        '404':
          description: Manifest not found
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/{id}/schema:
    put:
      summary: Update manifest schema (draft only)
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: ^[0-9a-z]{8,12}$
          description: Workspace ID (Base36, 8-12 characters)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                schema:
                  type: object
                  additionalProperties: true
                  description: Optional schema definition for the manifest
      responses:
        '200':
          description: Schema updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSetManifest'
        '400':
          description: Bad request or manifest not in draft status
        '404':
          description: Manifest not found
  /config/api/v1/projects/{projectId}/datasets/{dataSetId}/manifests/{id}/files/{fileId}:
    delete:
      summary: Delete a file from a manifest (draft only)
      tags:
      - Manifests
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: path
        name: dataSetId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: ^[0-9a-z]{8,12}$
          description: Workspace ID (Base36, 8-12 characters)
      - in: path
        name: fileId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: File deleted successfully
        '400':
          description: Bad request or manifest not in draft status
        '404':
          description: Manifest or file not found
...
components:
  schemas:
...
    DataSetManifest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dataSetId:
          type: string
          format: uuid
        manifestId:
          type: integer
          description: Monotonically increasing manifest version number
        status:
          type: string
          enum:
          - draft
          - committed
          - deprecated
          description: Manifest status - only draft manifests can be modified
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata for the manifest
        schema:
          type: object
          additionalProperties: true
          description: Optional schema definition for the manifest
        files:
          type: array
          items:
            $ref: '#/components/schemas/DataSetManifestFile'
          description: List of files in this manifest
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DataSetManifestFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        manifestId:
          type: string
          format: uuid
        fileName:
          type: string
          description: Original file name
        uri:
          type: string
          description: S3 URI or file path (e.g., s3://bucket/path/to/file)
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token obtained from Keycloak. You can get a token
        by logging into the GUI and copying the token from browser developer tools,
        or by using the OAuth2 flow below.
    oauth2:
      type: oauth2
      description: OAuth2 authentication via Keycloak. Click Authorize and login with
        your credentials.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.agentstudio.local:8443/realms/nemo/protocol/openid-connect/auth
          tokenUrl: https://auth.agentstudio.local:8443/realms/nemo/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email address